// src/app/models/transaction.ts
export interface transaction {
    checked: unknown;
    transactionId: number;
    transDate: Date;
    nama: string;
    noPatG: string;
    noResit: string;
    type: string;
    sekerap: number;
    lateks: number;
    weight: number;
    drc: number;
    total: number;
    unitPrice: number;

}